\name{ibex_hr}
\alias{ibex_hr}
\docType{data}
\title{
One year of heart rates of a capricorn free-living in the alps.
}
\description{
Heart rates were obtained using acoustic loggers in the rumen.
}
\usage{data("ibex_hr")}
\format{
  A data frame with 28454 observations on the following 2 variables.
  \describe{
    \item{\code{Time}}{datetime}
    \item{\code{Heartrate}}{in bpm}
  }
}

\references{
Signer, C., Ruf, T., & Arnold, W. (2011). Hypometabolism and basking: The strategies of Alpine ibex to endure harsh over‐wintering conditions. Functional Ecology, 25(3), 537-547.
}
\examples{
data(ibex_hr)
}
\keyword{datasets}
