\name{wb_year}
\alias{wb_year}
\docType{data}
\title{
One year of heart rates of a wild boar female.
}
\description{
Heart rates were obtained from DST centi- HRT, Star-Oddi, Gardabaer, Iceland.

}
\usage{data("wb_year")}
\format{
  A data frame with 3720 observations on the following 3 variables.
  \describe{
    \item{\code{x}}{x-axis, datetime}
    \item{\code{y}}{y-axis, heart rate}

  }
}

\references{
Ruf, T., Vetter, S. G., Painer, J., Stalder, G., & Bieber, C. (2021). Atypical for northern ungulates, energy metabolism is lowest during summer in female wild boars (Sus scrofa). Scientific Reports, 11(1), 1-12.

}
\examples{
data(wb_year)
}
\keyword{datasets}
